pyproject.toml
[tool.poetry]
name = "misc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
jinja2 = "^3.1.5"
pyyaml = "^6.0.2"
openai = "^1.60.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
